+++ ReleaseNotes.md
@@ -8,6 +8,7 @@
 * Make project names and paths in config file case insensitive (issue #63)
 * Simplify VSSDK package management by using the VSSDK meta package (issue #78)
 * Refresh configuration of current item when configuration file has changed (issue #76)
+* Add support for switching the solution configuration and solution platform (issue #70)
 
 ## Version 3.5
 * Fix support for configuration parameters of C++ projects (issues #55, #61)